changed simpair from uint32_t to uint64_t new revids are to long

+++ wikipedia-vis/ext/wikidb/include/contract.h
@@ -38,7 +38,7 @@ class Article : public Page {
 
     //member
     int4 words;
-    dbArray<uint32_t> comparisons;
+    dbArray<uint64_t> comparisons;
 
     TYPE_DESCRIPTOR((SUPERCLASS(Page), KEY(words, INDEXED|HASHED),
                     FIELD(comparisons)));
